Update app : wav dtype을 int16으로 변경

+++ src/app.py
@@ -45,7 +45,7 @@ def sound_pressure_level(signal):
 
 def record_audio(duration, fs):
     print("recording start")
-    recording = sd.rec((duration * fs) & 0xFFFFFFFF, samplerate=fs, channels=1, dtype='float64')
+    recording = sd.rec((duration * fs) & 0xFFFFFFFF, samplerate=fs, channels=1, dtype='int16')
     sd.wait()  # Wait until recording is finished
     print("recording done")
     return recording
@@ -57,7 +57,7 @@ def main():
 
     try:
         while True:
-            recording = sd.rec(fs & 0xFFFFFFFF, samplerate=fs, channels=1, dtype='float64')
+            recording = sd.rec(fs & 0xFFFFFFFF, samplerate=fs, channels=1, dtype='int16')
             sd.wait()
             spl = sound_pressure_level(recording[:,0])
             print(f"now: {spl:.2f} dB")
